budget: extract renderSankeySvg from hydrateCategorySankey (#599)
Move pure SVG layout and construction logic into sankey-render.ts so
hydrateCategorySankey drops below 200 lines and the renderer is
unit-testable in isolation. The hydrator retains event wiring,
hydration plumbing, and result mounting.
